% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_line_history}
\alias{add_line_history}
\title{Add line history}
\usage{
add_line_history(log)
}
\arguments{
\item{log}{An unested gitsum log, for example obtained through
\code{\link[=parse_log_detailed]{parse_log_detailed()}}.}
}
\description{
Add amount of lines added by commits (insertions - deletions) plus number of
total lines a file has according to past insertions and deletions
(via \code{\link[base:cumsum]{base::cumsum()}}.
}
\examples{
add_line_history(tidyr::unnest(gitsumlog))
}
