% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_n_times_changed_file}
\alias{add_n_times_changed_file}
\title{How many times did file get changed?}
\usage{
add_n_times_changed_file(log)
}
\arguments{
\item{log}{An unnested log.}
}
\description{
Add a variable \code{n_times_changed_file} to the unnested log, which contains
the number of times a file was changed according to \code{log}. This may be
helpful for subsetting.
}
