% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{combine_dir_and_base}
\alias{combine_dir_and_base}
\title{Derive full paths}
\usage{
combine_dir_and_base(dirs, base_names, extensions, reassignment_index)
}
\arguments{
\item{dirs}{Character vector with directories.}

\item{base_names}{List of pairs containing the old base name as the first
element and the new base name as the second.}

\item{reassignment_index}{Integer vector indicating the position of the
re-assignment in the original unnested detailed log.}
}
\description{
Derive full paths
}
\examples{
gitsum:::combine_dir_and_base(
  "R",
  list(c("a", "b"), c("API", "api2"), c("scr", "inst/include")),
  list("", "", "dplyr_types.h"), 1
)
}
