% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.R
\name{create_log}
\alias{create_log}
\title{Create the log raw data and clean up}
\usage{
create_log(path, file_name = ".log_stat.txt")
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}

\item{file_name}{the name of the temporary file.}
}
\description{
Create the log raw data and clean up
}
