% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{dump_parsed_log}
\alias{dump_parsed_log}
\alias{dump_last_commit}
\title{Dump a parsed log into .gitsum}
\usage{
dump_parsed_log(log, path = ".", over_write = FALSE)

dump_last_commit(log, path)
}
\arguments{
\item{log}{A parsed log as a tibble}

\item{path}{The path to the root directory of the gitsum repository.}
}
\description{
Store a parsed git log in a gitsum repository.
}
\section{Functions}{
\itemize{
\item \code{dump_last_commit}: Dumps the last commit into .gitsum
}}

