% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{ensure_enclosing}
\alias{ensure_enclosing}
\alias{ensure_curly_enclosing}
\alias{ensure_dash_enclosing}
\title{Ensure enclosing}
\usage{
ensure_curly_enclosing(raw_reassignment)

ensure_dash_enclosing(raw_reassignment)
}
\arguments{
\item{raw_assignment}{A character vector containing text from
raw assignemnts (see 'Examples').}
}
\description{
Make sure a string is enclosed with some values.
}
\section{Functions}{
\itemize{
\item \code{ensure_curly_enclosing}: Ensures curly braces enclosing.

\item \code{ensure_dash_enclosing}: Ensures dash enclosing.
}}

\examples{
gitsum:::ensure_curly_enclosing(
  c("API => api", "R/{a => b}")
)
gitsum:::ensure_leading_dash(
  c("{API = api}", "R/{a => b}",
  "vignettes/{ => notes}/mysql-setup.Rmd")
)
}
