% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ensure_gitusm_repo}
\alias{ensure_gitusm_repo}
\title{Make sure the repository is a gitsum repository and create one if it is not}
\usage{
ensure_gitusm_repo(path = ".")
}
\arguments{
\item{path}{The path to the root directory of the gitsum repository.}
}
\description{
Make sure the repository is a gitsum repository and create one if it is not
}
