% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.R
\name{extract_factory}
\alias{extract_factory}
\title{closure to extract regex pattern from vector}
\usage{
extract_factory(pattern)
}
\arguments{
\item{pattern}{The pattern the function should be able to extract.}
}
\description{
closure to extract regex pattern from vector
}
