% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_factories.R
\name{extract_factory_multiple}
\alias{extract_factory_multiple}
\title{closure to extract multiple regex pattern from vector}
\usage{
extract_factory_multiple(pattern)
}
\arguments{
\item{pattern}{The pattern the function should be able to extract.}
}
\description{
closure to extract multiple regex pattern from vector
}
\keyword{internal}
