% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{file_path}
\alias{file_path}
\title{Like \code{\link[base:file.path]{base::file.path()}}, but sorts our \code{character(1)} first}
\usage{
file_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors.}

\item{fsep}{the path separator to use.}
}
\description{
Like \code{\link[base:file.path]{base::file.path()}}, but sorts our \code{character(1)} first
}
