% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.R
\name{find_message_and_desc}
\alias{find_message_and_desc}
\title{extract message or description from a log}
\usage{
find_message_and_desc(raw, target)
}
\arguments{
\item{raw}{A character vector corresponding to one commit.}

\item{target}{either "message" or "description}
}
\description{
For message and description extraction, a different approach is used and
  one cannot rely on the approach with the simple extraction factory.
}
