% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.R
\name{get_log_regex}
\alias{get_log_regex}
\title{get log of a github repo via regex}
\usage{
get_log_regex(path = ".")
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}
}
\description{
get log of a github repo via regex
}
