% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{get_log_simple}
\alias{get_log_simple}
\title{get the log from a git repo}
\usage{
get_log_simple(path = ".")
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}
}
\description{
get the log from a git repo
}
