% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.R
\name{get_pattern}
\alias{get_pattern}
\title{regex patterns for extraction}
\usage{
get_pattern()
}
\value{
returns a named vector with regex patterns to extract from a
git log.
}
\description{
regex patterns for extraction
}
