% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get_pattern_multiple}
\alias{get_pattern_multiple}
\title{regex patterns for extraction}
\usage{
get_pattern_multiple()
}
\value{
returns a named vector with regex patterns to extract from a
git log.
}
\description{
regex patterns for extraction
}
\keyword{internal}
