% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_log.R
\name{get_raw_log}
\alias{get_raw_log}
\title{Obtain the log raw data}
\usage{
get_raw_log(path, file_name = NULL, remove = is.null(file_name),
  commit_range = NULL)
}
\arguments{
\item{path}{The path to the git directory one wants to create summaries for.}

\item{file_name}{The name of the temporary file. If \code{NULL}, a file is
created, otherwise, a file is read.}

\item{remove}{Whether a log should be deleted after read in.}

\item{commit_range}{A string of the form "hash1..hash2" indicating the commit
range to parse. \code{NULL} means all commits.}
}
\description{
Obtain the log raw data
}
