% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{git_log_simple}
\alias{git_log_simple}
\title{Obtain a simple log for a git repo}
\usage{
git_log_simple(path = ".")
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}
}
\description{
This function returns a git log in a tabular format.
}
\seealso{
See \link{git_log_detailed} for a slower alternative with more information.
}
