% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{git_report}
\alias{git_report}
\title{Summarize a git repo}
\usage{
git_report(path = ".", output_file = NULL,
  output_format = "html_document", template = "gitsum::repo_summary_simple",
  input_file = NULL, directory = "gitsum", cached = TRUE,
  .libpath_index = 1)
}
\arguments{
\item{path}{The path to the git directory one wants to create summaries for.}

\item{output_file}{Output file. If \code{NULL} then a default based on
  the name of the input file is chosen.}

\item{output_format}{R Markdown output format to convert to. Pass
  \code{"all"} to render all formats defined within the file. Pass
  the name of a format (e.g. \code{"html_document"}) to render a single
  format or pass a vector of format names to render multiple formats.
  Alternatively you can pass an output format object;
  e.g. \code{html_document()}. If \code{NULL} is passed then the
  output format is the first one defined in the YAML metadata of the
  input file (defaulting to HTML if none is specified).}

\item{template}{the template file to be used. Either one from a package
(e.g. \code{gitsum::repo_summary_simple}), or a user defined template,
which is then indicated with a path to this file including extension, i.e.
\code{/users/xzy/a_template.Rmd}.}

\item{input_file}{the name of the temporary file that contains log data on
which the report is base on. If \code{NULL}, a log file is created according to
\code{path}, otherwise, a file is read.}

\item{directory}{A directory to which the reports should be stored.}

\item{cached}{Whether or not the report should be cached with rmarkdown, so
the graphs are writen to a file.}

\item{.libpath_index}{index of the Library according to \link[base:libPaths]{base::libPaths} in
which the template should be searched. The list with the available
directories is returned invisibly for situations in which \code{.libPaths()}
cannot be emulated.}
}
\description{
Producing a summary report of a repo
}
\details{
Other packages can define templates. They must be store an .Rmd
template in inst/report_templates. The data passed to the report template
is log, the parsed log table.
}
