% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{group_reassignment}
\alias{group_reassignment}
\title{Create indices for splitting \code{log} into a group to update and a group to
not update}
\usage{
group_reassignment(log, transition)
}
\arguments{
\item{log}{An unnested detailed log.}

\item{transition}{A named list with one entry from the transition model.}
}
\description{
Create indices for splitting \code{log} into a group to update and a group to
not update
}
