% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is_detailed_log}
\alias{is_detailed_log}
\title{Check whether an object is a detailed log}
\usage{
is_detailed_log(log, nested = NA)
}
\arguments{
\item{log}{The object to inspect.}

\item{nested}{Whether or not the object to inspect should be nested or not.
\code{NA} indicates that both nested and unnested logs shoud pass the
inspection.}
}
\description{
Check whether an object is a detailed log
}
\keyword{internal}
