% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{is_name_change}
\alias{is_name_change}
\title{Detetct whether a \code{changed_file}-entry is a renaming}
\usage{
is_name_change(changed_file,
  reassignment_pattern = "\\\\{.+\\\\s\\\\=\\\\>.+\\\\}$")
}
\arguments{
\item{changed_file}{The column \code{changed_file} from a parsed unnested
detailed log.}

\item{reassignment_pattern}{The regular expression pattern that corresponds
to renaming.}
}
\description{
Detetct whether a \code{changed_file}-entry is a renaming
}
\examples{
gitsum:::is_name_change("R/{gitsum.R => gitsum-package.R}")
}
