% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{manage_gitsum}
\alias{manage_gitsum}
\alias{init_gitsum}
\alias{remove_gitsum}
\alias{init_gitsum}
\alias{remove_gitsum}
\title{Manage a gitsum repository}
\usage{
init_gitsum(path = ".", over_write = FALSE)

remove_gitsum(path)
}
\arguments{
\item{path}{The path to the repository to initialize.}

\item{over_write}{Whether or not an existing directory should be overwritten.}
}
\description{
A gitsum repository is a git repository with an additional .gitsum folder in
the root directory that is mainly used to store a parsed git log as a tibble.
}
\section{Functions}{
\itemize{
\item \code{init_gitsum}: Calling this function will parse the available git
history and dump it into the directory .gitsum.

\item \code{remove_gitsum}: Removes the gitsum repository, i.e just the folder
.gitsum and its contents.
}}

