% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_log}
\alias{nest_log}
\alias{unnest_log}
\title{Nest/unnest a gitsum log}
\usage{
nest_log(log)

unnest_log(log)
}
\arguments{
\item{log}{A log to nest/unnest.}
}
\description{
Forwards to \code{\link[tidyr:nest]{tidyr::nest()}} while taking care of which columns have to be
nested.

Forwards to \code{\link[tidyr:unnest]{tidyr::unnest()}} while taking care of which columns have to be
nested unnested.
}
