% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_lines}
\alias{parse_lines}
\title{Turn a raw log of lines into a tabular format}
\usage{
parse_lines(lines)
}
\arguments{
\item{lines}{The output of \code{\link[=get_raw_log]{get_raw_log()}}.}
}
\description{
Turn a raw log of lines into a tabular format
}
\keyword{internal}
