% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.R
\name{parse_log}
\alias{parse_log}
\title{parse a raw log}
\usage{
parse_log(raw, fnc_list)
}
\arguments{
\item{raw}{a list of character vectors, each list element corresponding to
one commit.}

\item{fnc_list}{the list list of named functions that return the elements
we want to extract from the log, i.e. author, hash etc.}
}
\description{
parse a raw log
}
