% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{parse_reassignment}
\alias{parse_reassignment}
\title{Parse a raw reassignment into a list}
\usage{
parse_reassignment(raw_reassignment, reassignment_index)
}
\arguments{
\item{raw_reassignment}{Character vector with raw re-assignments.}

\item{reassignment_index}{Integer vector indicating the position of the
re-assignment in the original unnested detailed log.}
}
\description{
Takes a vector of raw reassignments and turns them into a list of elements.
Every element contains the old path, the new path and the index of the
reassignment.
}
\examples{
gitsum:::parse_reassignment(
  c("R/{a => b}", "API => api2", "{src => inst/include}/dplyr_types.h"),
  c(1, 2, 33)
)
}
\keyword{internal}
