% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_gitsum}
\alias{read_gitsum}
\alias{read_log}
\alias{read_last_commit}
\alias{read_last_hash}
\title{Read gitsum data into R}
\usage{
read_log(path = ".")

read_last_commit(path = ".")

read_last_hash(path = ".")
}
\arguments{
\item{path}{The path to the root directory of the gitsum repository.}
}
\description{
Read gitsum data into R
}
\section{Functions}{
\itemize{
\item \code{read_log}: Reads a parsed log.

\item \code{read_last_commit}: Reads the last parsed commit.

\item \code{read_last_hash}: Reads the last parsed hash.
}}

\keyword{internal}
