% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_git}
\alias{report_git}
\title{Create a summary report of a git repository}
\usage{
report_git(path = ".", output_file = NULL,
  output_format = "html_document",
  template = "gitsum::repo_summary_simple", input_file = NULL,
  directory = "gitsum", cached = TRUE, .libpath_index = 1)
}
\arguments{
\item{path}{The path to the git directory one wants to create summaries for.}

\item{output_file}{The name of the output file. If using \code{NULL} then the
output filename will be based on filename for the input file. If a filename
is provided, a path to the output file can also be provided. Note that the
\code{output_dir} option allows for specifying the output file path as well,
however, if also specifying the path, the directory must exist. If
\code{output_file} is specified but does not have a file extension, an
extension will be automatically added according to the output format. To
avoid the automatic file extension, put the \code{output_file} value in
\code{\link{I}()}, e.g., \code{I('my-output')}.}

\item{output_format}{The R Markdown output format to convert to. The option
\code{"all"} will render all formats defined within the file. The option can
be the name of a format (e.g. \code{"html_document"}) and that will render
the document to that single format. One can also use a vector of format
names to render to multiple formats. Alternatively, you can pass an output
format object (e.g. \code{html_document()}). If using \code{NULL} then the
output format is the first one defined in the YAML frontmatter in the input
file (this defaults to HTML if no format is specified there).}

\item{template}{the template file to be used. Either one from a package
(e.g. \code{gitsum::repo_summary_simple}), or a user defined template,
which is then indicated with a path to this file including extension, i.e.
\code{/users/xzy/a_template.Rmd}.}

\item{input_file}{the name of the temporary file that contains log data on
which the report is base on. If \code{NULL}, a log file is created according to
\code{path}, otherwise, a file is read.}

\item{directory}{A directory to which the reports should be stored.}

\item{cached}{Whether or not the report should be cached with rmarkdown, so
the graphs are written to a file.}

\item{.libpath_index}{index of the Library according to \link[base:libPaths]{base::libPaths} in
which the template should be searched. The list with the available
directories is returned invisibly for situations in which \code{.libPaths()}
cannot be emulated.}
}
\description{
Producing a summary report of a repo.
}
\details{
Other packages can define templates. They must be store an .Rmd
template in inst/report_templates. The data passed to the report template
is log, the parsed log table.
}
