% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{rmd_simple}
\alias{rmd_simple}
\title{Get a RMarkdown summary document of a repo}
\usage{
rmd_simple(path = ".", output_file = NULL, output_format = "all",
  template = "gitsum::repo_summary_simple", directory = "gitsum",
  cached = FALSE, .libpath_index = 1)
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}

\item{output_file}{Output file. If \code{NULL} then a default based on
  the name of the input file is chosen.}

\item{output_format}{R Markdown output format to convert to. Pass
  \code{"all"} to render all formats defined within the file. Pass
  the name of a format (e.g. \code{"html_document"}) to render a single
  format or pass a vector of format names to render multiple formats.
  Alternatively you can pass an output format object;
  e.g. \code{html_document()}. If \code{NULL} is passed then the
  output format is the first one defined in the YAML metadata of the
  input file (defaulting to HTML if none is specified).}

\item{template}{the template file to be used. Either one from a package
(e.g. \code{gitsum::repo_summary_simple}), or a user defined template,
which is then indicated with a path to this file including extension, i.e.
\code{/users/xzy/a_template.Rmd}.}

\item{directory}{A directory to which the reports should be stored.}

\item{.libpath_index}{index of the Library according to \link{base:.libPaths} in
which the template should be searched. The list with the available
directories is returned invisibly for situations in which `.libPaths()``
cannot be emulated.}

\item{cache}{Whether or not the report should be cached with rmarkdown, so
the graphs are writen to a file.}

\item{file_out}{The path to the file the report should be created}
}
\description{
Get a RMarkdown summary document of a repo
}
\details{
Other packages can define templates. They must be store an .Rmd
template in inst/report_templates. The data passed to the report template
is log, the parsed log table.
}
