% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{separate_dir_and_reassignment}
\alias{separate_dir_and_reassignment}
\title{Separate the directory from the reassignment}
\usage{
separate_dir_and_reassignment(raw_reassignment)
}
\description{
Separate the directory from the reassignment
}
\examples{
separate_dir_and_reassignment("R/{gitsum.R => gitsum-package.R}")
}
