% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{set_changed_file_to_latest_name}
\alias{set_changed_file_to_latest_name}
\title{Set the column  \code{changed_files} of a \code{log} to the latest name of the file}
\usage{
set_changed_file_to_latest_name(log)
}
\arguments{
\item{log}{An unnested detailed log.}
}
\description{
\code{\link[=parse_log_detailed]{parse_log_detailed()}} parses a git log. However, changed file names
are not corrected retroactively. Hence, if one want's to process file-specific
commit data for one file accross different names for that file, the column
\code{changed_files} has to be unified. This is done with this function by
updating the file names recursively and setting them to the latest name used.
}
\examples{
library("magrittr")
gitsumlog \%>\%
  tidyr::unnest() \%>\%
  set_changed_file_to_latest_name() \%>\%
  add_line_history()
}
