% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{update_changed_file}
\alias{update_changed_file}
\title{Apply one file name update to a log}
\usage{
update_changed_file(log, transition)
}
\arguments{
\item{log}{An unnested detailed log.}

\item{transition}{A named list with one entry from the transition model.}
}
\description{
Apply one file name update to a log
}
