% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{update_changed_file_sequentially}
\alias{update_changed_file_sequentially}
\title{Apply all updates of file names sequentially to a log}
\usage{
update_changed_file_sequentially(log, transition_model)
}
\arguments{
\item{log}{An unnested detailed log.}

\item{transition_model}{A transition model obtained from
\code{\link[=parse_reassignment]{parse_reassignment()}}.}
}
\description{
Recursively updates the column \code{changed_file} of \code{log} according to a
transition scheme provided in \code{transition_model}.
}
