% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineer-features-base.R
\name{add_attributes_detailed}
\alias{add_attributes_detailed}
\title{Add base attributes to a log}
\usage{
add_attributes_detailed(log)
}
\description{
Derives some attributes from the parsed raw log and converts them to the
desired format.
The following attributes merit special mentioning:
}
\details{
\itemize{
\item commit_nr: The number of the commit in the repo. This can be used to sort
commits (in contrast to the commit date) in chronological order, that is,
in the oder in which they were committed.
}
}
