% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineer-features-base.R
\name{add_is_exact}
\alias{add_is_exact}
\title{Create a column indicating whether insertions and deletions are exact counts}
\usage{
add_is_exact(log)
}
\description{
Create a column indicating whether insertions and deletions are exact counts
}
