% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{add_line_history}
\alias{add_line_history}
\title{Add line history}
\usage{
add_line_history(log)
}
\arguments{
\item{log}{A log.}
}
\description{
Add amount of lines added by commits (insertions - deletions) plus
number of total lines a file has according to past insertions and
deletions (via \code{\link[base:cumsum]{base::cumsum()}}.
}
