% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{combine_dir_and_base_one}
\alias{combine_dir_and_base_one}
\title{Combine dir with old and new basename}
\usage{
combine_dir_and_base_one(dirname, basename, reassignment_index)
}
\description{
Takes a dir and the new
}
