% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{expect_class}
\alias{expect_class}
\title{Class validation}
\usage{
expect_class(data, class_mapping, must_check_all)
}
\arguments{
\item{data}{A data frame to check.}

\item{class_mapping}{A data frame that contains two columns: name and class,
whereas the values in name indicate the name of a column to check, class
indicates the target class of the column.}

\item{must_check_all}{Whether or not all columns need to be checked in order
to avoid a warning.}
}
\description{
Check whether columns of a data frame match are of a certain class and return
an informative warning otherwise.
}
\examples{
library(tibble)
library(magrittr)
class_mapping <- tribble(
  ~ name,  ~ class,
  "cyl", "integer",
  "model", "character"
)
rownames_to_column(mtcars, var= "model") \%>\%
  expect_class(class_mapping, must_check_all = FALSE)
}
