% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed.R
\name{git_log_detailed}
\alias{git_log_detailed}
\title{get log of a github repo via regex}
\usage{
git_log_detailed(path = ".", file_name = NULL)
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}

\item{file_name}{the name of the temporary file. If \code{NULL}, a file is created,
otherwise, a file is read.}
}
\description{
get log of a github repo via regex
}
