% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitsum-package.R
\docType{package}
\name{gitsum}
\alias{gitsum}
\alias{gitsum-package}
\alias{gitsum-package}
\title{Parse and summarize a git repository}
\description{
\code{gitsum} provides tools to parse a git repository history and
can create automated markdown reports from the parsed history.
}
\author{
Lorenz Walthert \email{lorenz.walthert@icloud.com}, \cr
}
