% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed.R
\name{parse_log_detailed}
\alias{parse_log_detailed}
\title{Obtain a detailed git log}
\usage{
parse_log_detailed(path = ".", na_to_zero = TRUE, file_name = NULL)
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}

\item{file_name}{the name of the temporary file. If \code{NULL}, a file is
created, otherwise, a file is read.}
}
\value{
A parsed git log as a nested tibble. Each line corresponds to a
commit. The unnested column names are: \cr
short_hash, author_name, date, short_message, hash, left_parent,
right_parent, author_email, weekday, month, monthday, time, year, timezone,
message, description, total_files_changed, total_insertions,
total_deletions, short_description, is_merge \cr
The nested columns contain more information on each commit. The column
names are: \cr
changed_file, edits, insertions, deletions.
}
\description{
This function returns a git log in a tabular format.
}
\details{
\itemize{
\item Note that for merge commmits, the following columns are \code{NA} if
the opotion \code{na_to_zero} is set to \code{FALSE}.:
total_files_changed, total_insertions, total_deletions, changed_file,
edits, deletions, insertions.
\item Note that for binary files, the following columns are 0: edits, deletions,
insertions.
}
}
\section{Warning}{

The number of edits, insertions, and deletions (on a file level) are based
on \code{git log --stat} and the number of \code{+} and \code{-} in this log. The number
of \code{+} and \code{-} may not sum up to the edits indicated as a scalar after "|"
for commits with very many changed lines since for those, the \code{+} and \code{-}
only indicate the relavite share of insertinos and edits. Therefore,
\code{parse_log_detailed()} normalizes the insertions and deletions and rounds
these after the normalization to achieve more consistent results. However,
there is no guarantee that these numbers are always exact. The column
is_exact indicates for each changed file within a commit wether the result
is exact (which is the case if the normalizing constant was one).
}

\seealso{
See \link{parse_log_simple} for a fast alternative with less information.
}
