% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_log_one}
\alias{parse_log_one}
\title{parse a raw log}
\usage{
parse_log_one(raw, fnc_list, has_merge)
}
\arguments{
\item{raw}{a character vector corresponding to one commit.}

\item{fnc_list}{the list of named functions that return the elements
we want to extract from the log, i.e. author, hash etc.}

\item{has_merge}{Whether or not \code{raw} is a merge commit. This is relevant
because merge commits have additional lines.}
}
\description{
parse a raw log
}
\keyword{internal}
