% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{parse_log_simple}
\alias{parse_log_simple}
\title{Obtain a simple log for a git repo}
\usage{
parse_log_simple(path = ".", file_name = NULL)
}
\arguments{
\item{path}{the path to the git directory one wants to create summaries for.}

\item{file_name}{the name of the temporary file. If \code{NULL}, a file is
created, otherwise, a file is read.}
}
\description{
This function returns a git log in a tabular format.
}
\seealso{
See \link{parse_log_detailed} for a slower alternative with more
information.
}
