% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed-files.R
\name{update_changed_file_one}
\alias{update_changed_file_one}
\title{Apply one file name update to an unnested log that contains commits
up to that name change. Also needs to update the last entry in log, which
is the transition itself.}
\usage{
update_changed_file_one(log, transition)
}
\description{
Apply one file name update to an unnested log that contains commits
up to that name change. Also needs to update the last entry in log, which
is the transition itself.
}
